+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Some of these are pending, some deserve PRs, others are just minor tweaks
 
 * Delete extensions (`mworzala/Minestom` @ `no_more_extensions`)
 * Block face in digging events (`mworzala/Minestom` @ `block_break_face`)
+* Add cursor position to block place and neighbor updates (`Moulberry/Minestom` @ `block_placement_rewrite_2`)